Syntax Reference

CSTParser.check_reformat

CSTParser.check_span

CSTParser.compare

CSTParser.contributes_scope

CSTParser.get_id

CSTParser.get_sig

CSTParser.lex_ws_comment

CSTParser.parse

CSTParser.parse_array

CSTParser.parse_block

CSTParser.parse_call

CSTParser.parse_comma_sep

CSTParser.parse_compound

CSTParser.parse_curly

CSTParser.parse_doc

CSTParser.parse_dot_mod

CSTParser.parse_expression

CSTParser.parse_generator

CSTParser.parse_if

CSTParser.parse_iterator

CSTParser.parse_iterators

CSTParser.parse_macrocall

CSTParser.parse_parameters

CSTParser.parse_paren

CSTParser.parse_ref

CSTParser.parse_string_or_cmd

CSTParser.parse_tuple

CSTParser.parse_unary

CSTParser.remlineinfo!

CSTParser.@closer

CSTParser.@default

CSTParser.@nocloser

CSTParser.@precedence

Main

CSTParser.parse_expression — Method

parse_expression(ps)

Parses an expression until closer(ps) == true. Expects to enter the ParseState the token before the the beginning of the expression and ends on the last token.

Acceptable starting tokens are:

A keyword

An opening parentheses or brace.

An operator.

An instance (e.g. identifier, number, etc.)

An @.

CSTParser.parse — Function

parse(str, cont = false)

Parses the passed string. If cont is true then will continue parsing until the end of the string returning the resulting expressions in a TOPLEVEL block.

CSTParser.parse_compound — Method

parse_compound(ps::ParseState, ret::EXPR)

Attempts to parse a compound expression given the preceding expression ret.

CSTParser.parse_doc — Method

Used for top-level parsing - attaches documentation (such as this) to expressions.

CSTParser.parse_paren — Method

parse_paren(ps, ret)

Parses an expression starting with a (.

CSTParser.remlineinfo! — Method

remlineinfo!(x)

Removes line info expressions. (i.e. Expr(:line, 1))

CSTParser.contributes_scope — Method

contributes_scope(x)

Checks whether the body of x is included in the toplevel namespace.

CSTParser.get_id — Method

get_id(x)

Get the IDENTIFIER name of a variable, possibly in the presence of type declaration operators.

CSTParser.get_sig — Method

get_sig(x)

Returns the full signature of function, macro and datatype definitions. Should only be called when has_sig(x) == true.

CSTParser.lex_ws_comment — Method

lex_ws_comment(l::Lexer, c)

Having hit an initial whitespace/comment/semicolon continues collecting similar Chars until they end. Returns a WS token with an indication of newlines/ semicolons. Indicating a semicolons takes precedence over line breaks as the former is equivalent to the former in most cases.

CSTParser.check_reformat — Method

check_reformat()

Reads and parses all files in current directory, applys formatting fixes and checks that the output AST remains the same.

CSTParser.check_span — Function

check_span(x, neq = [])

Recursively checks whether the span of an expression equals the sum of the span of its components. Returns a vector of failing expressions.

CSTParser.compare — Method

compare(x,y)

Recursively checks whether two Base.Expr are the same. Returns unequal sub- expressions.

CSTParser.@closer — Macro

@closer ps rule body

Continues parsing closing on rule.

CSTParser.@default — Macro

@default ps body

Parses the next expression using default closure rules.

CSTParser.@nocloser — Macro

@nocloser ps rule body

Continues parsing not closing on rule.

CSTParser.@precedence — Macro

@precedence ps prec body

Continues parsing binary operators until it hits a more loosely binding operator (with precdence lower than prec).

Components

CSTParser.parse_block — Function

Continue parsing statements until an element of closers is hit (usually end). Statements are grouped in a Block EXPR.

CSTParser.parse_call — Function

parse_call(ps, ret)

Parses a function call. Expects to start before the opening parentheses and is passed the expression declaring the function name, ret.

CSTParser.parse_comma_sep — Function

Parses a comma separated list, optionally allowing for conversion of assignment (=) expressions to Kw.

CSTParser.parse_dot_mod — Function

Helper function for parsing import/using statements.

CSTParser.parse_generator — Method

parse_generator(ps)

Having hit for not at the beginning of an expression return a generator. Comprehensions are parsed as SQUAREs containing a generator.

CSTParser.parse_iterator — Method

Parses an iterator, allowing for the preceding keyword outer. Returns an error expression if an invalid expression is parsed (anything other than =, in, ∈).

CSTParser.parse_iterators — Function

Parses a group of iterators e.g. used in a for loop or generator. Can allow for a succeeding Filter expression.

CSTParser.parse_macrocall — Method

parse_macrocall(ps)

Parses a macro call. Expects to start on the @.

CSTParser.parse_parameters — Function

Parses parameter arguments for a function call (e.g. following a semicolon).

CSTParser.parse_if — Function

parse_if(ps, ret, nested=false, puncs=[])

Parse an if block.

CSTParser.parse_array — Function

parse_array(ps)

Having hit '[' return either:

A vect

A vcat

A comprehension

An array (vcat of hcats)

CSTParser.parse_curly — Method

parse_curly(ps, ret)

Parses the juxtaposition of ret with an opening brace. Parses a comma seperated list.

CSTParser.parse_ref — Method

parse_ref(ps, ret)

Handles cases where an expression - ret - is followed by [. Parses the following bracketed expression and modifies it's .head appropriately.

CSTParser.parse_tuple — Function

parse_tuple(ps, ret)

ret is followed by a comma so tries to parse the rest of the tuple.

CSTParser.parse_unary — Method

parse_unary(ps)

Having hit a unary operator at the start of an expression return a call.

CSTParser.parse_string_or_cmd — Function

parsestringor_cmd(ps)

When trying to make an INSTANCE from a string token we must check for interpolating opoerators.